minor, provide resz() (similar to crea()) to resize an rbm
We remap the old (still active) rbm map to a MAP_PRIVATE mapping, then
ftruncate() the whole file to meet the new size criteria (MAP_SHARED),
then copy from the private to the shared map, and deactivate the private
mapping.

If there's new neurons (new size > old size) the new neurons will be
initialised with the same wobbling that created the old network.

Signed-off-by: Sebastian Freundt <freundt@ga-group.nl>